+++ proto/prysm/v2/BUILD.bazel
@@ -31,6 +31,7 @@ proto_library(
     deps = [
         "//proto/eth/ext:proto",
         "//proto/prysm/v1alpha1:proto",
+        "//proto/beacon/rpc/v1:v1_proto",
         "@com_github_grpc_ecosystem_grpc_gateway_v2//protoc-gen-openapiv2/options:options_proto",
         "@com_google_protobuf//:any_proto",
         "@com_google_protobuf//:descriptor_proto",
@@ -74,6 +75,7 @@ go_proto_library(
     deps = [
         "//proto/eth/ext:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
+        "//proto/beacon/rpc/v1:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_grpc_ecosystem_grpc_gateway_v2//protoc-gen-openapiv2/options:options_go_proto",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
@@ -100,6 +102,7 @@ go_proto_library(
     deps = [
         "//proto/eth/ext:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
+        "//proto/beacon/rpc/v1:go_default_library",
         "@com_github_grpc_ecosystem_grpc_gateway_v2//protoc-gen-openapiv2/options:options_go_proto",
         "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
         "@go_googleapis//google/api:annotations_go_proto",
@@ -125,6 +128,7 @@ go_library(
     deps = SSZ_DEPS + [
         "//proto/eth/ext:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
+        "//proto/beacon/rpc/v1:go_default_library",
         "@com_github_grpc_ecosystem_grpc_gateway_v2//runtime:go_default_library",
         "@com_github_grpc_ecosystem_grpc_gateway_v2//utilities:go_default_library",
         "@io_bazel_rules_go//proto/wkt:empty_go_proto",
