--- proto/validator/accounts/v2/BUILD.bazel
@@ -1,70 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library")
-load("@rules_proto//proto:defs.bzl", "proto_library")
-load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
-
-# gazelle:ignore
-proto_library(
-    name = "ethereum_validator_accounts_v2_proto",
-    srcs = [
-        "keymanager.proto",
-        "web_api.proto",
-    ],
-    visibility = ["//visibility:public"],
-    deps = [
-        "//proto/beacon/rpc/v1:v1_proto",
-        "//proto/prysm/v1alpha1:proto",
-        "//proto/prysm/v2:proto",
-        "//proto/eth/ext:proto",
-        "@com_google_protobuf//:empty_proto",
-        "@go_googleapis//google/api:annotations_proto",
-    ],
-)
-
-go_proto_library(
-    name = "ethereum_validator_account_go_proto",
-    compilers = [
-        "@prysm//:cast_grpc_proto_compiler",
-    ],
-    importpath = "github.com/prysmaticlabs/prysm/proto/validator/accounts/v2",
-    proto = ":ethereum_validator_accounts_v2_proto",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//proto/beacon/rpc/v1:go_default_library",
-        "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_github_prysmaticlabs_eth2_types//:go_default_library",
-        "//proto/prysm/v1alpha1:go_default_library",
-        "//proto/prysm/v2:go_default_library",
-        "//proto/eth/ext:go_default_library",
-        "@go_googleapis//google/api:annotations_go_proto",
-        "@io_bazel_rules_go//proto/wkt:empty_go_proto",
-    ],
-)
-
-go_proto_library(
-    name = "ethereum_validator_account_gateway_proto",
-    compilers = [
-        "@prysm//:grpc_gateway_proto_compiler",
-    ],
-    embed =["ethereum_validator_account_go_proto"],
-    importpath = "github.com/prysmaticlabs/prysm/proto/validator/accounts/v2",
-    proto = ":ethereum_validator_accounts_v2_proto",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//proto/beacon/rpc/v1:go_default_library",
-        "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
-        "@io_bazel_rules_go//proto/wkt:empty_go_proto",
-        "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_github_prysmaticlabs_eth2_types//:go_default_library",
-        "//proto/prysm/v1alpha1:go_default_library",
-        "//proto/prysm/v2:go_default_library",
-	    "//proto/eth/ext:go_default_library",
-        "@go_googleapis//google/api:annotations_go_proto",
-    ],
-)
-
-go_library(
-    name = "go_default_library",
-    embed = [":ethereum_validator_account_gateway_proto"],
-    importpath = "github.com/prysmaticlabs/prysm/proto/validator/accounts/v2",
-    visibility = ["//visibility:public"],
-)
