+++ validator/rpc/server_test.go
@@ -1,7 +1,7 @@
 package rpc
 
 import (
-	pb "github.com/prysmaticlabs/prysm/proto/validator/accounts/v2"
+	pb "github.com/prysmaticlabs/prysm/proto/prysm/v2"
 )
 
 var _ pb.AuthServer = (*Server)(nil)
